+++ dolibarr-default/lib/sample.lib.php
@@ -8,19 +8,18 @@
 
  This program is free software; you can redistribute it and/or modify it under the terms of the GNU General Public License as published by the Free Software Foundation; either version 3 of the License, or (at your option) any later version.
 
- This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details.
+ This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the IMPLIED WARRANTY of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details.
 
  You should have received a copy of the GNU General Public License (./LICENSE.md).
 
  # FILE INFORMATION
-	DEFGROUP:
-	INGROUP:
-	REPO:
-	FILE:
-	VERSION:
-	BRIEF:
-	PATH:
-	NOTE:
+ DEFGROUP: <ProjectGroup>
+ INGROUP: <Subgroup>
+ REPO: <RepositoryURL>
+ VERSION: <Version>
+ PATH: <RelativePath>
+ BRIEF: <ShortDescription>
+ NOTE: <OptionalNotes>
 */
 /**
  * Prepare admin pages header